Common ReactionEvent interface
Added an interface for ReactionAddEvent and ReactionRemoveEvent to remove code duplication and allow better Generics
